requirements.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.6
Django==2.2.4
django-cors-headers==3.0.2
django-extensions==2.2.1
django-filter==2.2.0
django-rest-framework==0.1.0
django-rest-swagger==2.1.2
djangorestframework==3.10.2
djongo==1.2.33
idna==2.8
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
pycodestyle==2.5.0
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
pymongo==3.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.2.4
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
virtualenv==16.7.2
wrapt==1.11.2
